implement: extract CLI bootstrap boundary
